separate api into another app, another app use a different port for testing
